move `src/vs/workbench/services/themes/test/node/tokenStyleResolving.test.ts` into `node` layer
https://github.com/microsoft/vscode/issues/171508
