+++ src/vs/platform/extensionResourceLoader/common/extensionResourceLoaderService.ts
