+++ src/vs/workbench/services/themes/test/node/color-theme.json
